# swagger.yaml
openapi: 3.0.0
info:
  title: Wallet Service API
  version: 1.0.0
servers:
  - url: http://localhost:3000

paths:
  /:
    get:
      summary: API home
      responses:
        "200":
          description: OK

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK

  /wallets:
    get:
      summary: List wallets
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: OK
    post:
      summary: Create wallet
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                currency:
                  type: string
                  example: EGP
      responses:
        "201":
          description: Created

  /wallets/{id}:
    get:
      summary: Get wallet by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Wallet not found

  /wallets/{id}/ledger:
    get:
      summary: Get wallet ledger (history)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: OK
        "404":
          description: Wallet not found

  /wallets/{id}/deposit:
    post:
      summary: Deposit into wallet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_cents]
              properties:
                amount_cents:
                  type: integer
                  example: 10000
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
        "404":
          description: Wallet not found

  /wallets/{id}/withdraw:
    post:
      summary: Withdraw from wallet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_cents]
              properties:
                amount_cents:
                  type: integer
                  example: 2500
      responses:
        "200":
          description: OK
        "400":
          description: Insufficient funds / validation error
        "404":
          description: Wallet not found

  /transfers:
    get:
      summary: List transfers
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: OK

    post:
      summary: Create transfer
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            example: 11111111-1111-1111-1111-111111111111
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_wallet_id, to_wallet_id, amount_cents]
              properties:
                from_wallet_id:
                  type: string
                to_wallet_id:
                  type: string
                amount_cents:
                  type: integer
                  example: 1000
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
        "409":
          description: Idempotency key reused with different payload
        "404":
          description: Wallet not found

  /transfers/{id}:
    get:
      summary: Get transfer by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: Transfer not found